+++ lib/src/wa_server.dart
@@ -248,5 +248,5 @@ class WaServer {
 /// A class that holds version information for the server.
 class _Info {
   /// The version of the server.
-  final String version = '1.1.14';
+  final String version = '1.1.15';
 }